# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MARSA Yacht Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /yachts:
    get:
      operationId: listYachts
      tags: [yachts]
      summary: List live yachts
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: categoryId
          in: query
          schema: { type: string }
        - name: locationId
          in: query
          schema: { type: string }
        - name: intent
          in: query
          schema:
            type: string
            enum: [rent, buy]
            default: rent
        - name: capacity
          in: query
          schema: { type: integer }
        - name: date
          in: query
          description: Filter by availability date (YYYY-MM-DD)
          schema: { type: string }
        - name: templateId
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          description: Minimum price in EGP (inclusive)
          schema: { type: number }
        - name: maxPrice
          in: query
          description: Maximum price in EGP (inclusive)
          schema: { type: number }
        - name: features
          in: query
          description: Comma-separated list of required features e.g. "wifi,ac"
          schema: { type: string }
      responses:
        "200":
          description: List of yachts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YachtListResponse"
  /yachts/{id}:
    get:
      operationId: getYacht
      tags: [yachts]
      summary: Get yacht details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Yacht detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YachtDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /yachts/{id}/slots:
    get:
      operationId: getYachtSlots
      tags: [yachts]
      summary: Get available slots for a yacht (path-param version)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: from
          in: query
          required: true
          schema: { type: string }
        - name: to
          in: query
          required: true
          schema: { type: string }
        - name: templateId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityListResponse"
  /yachts/availability:
    get:
      operationId: getYachtAvailability
      tags: [yachts]
      summary: Get available slots for a yacht
      parameters:
        - name: yachtId
          in: query
          required: true
          schema: { type: string }
        - name: from
          in: query
          required: true
          schema: { type: string }
        - name: to
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Users ──────────────────────────────────────────────────────────────────
...
    Yacht:
      type: object
      properties:
        id:
          type: string
        hostId:
          type: string
        categoryId:
          type: ["string", "null"]
        locationId:
          type: ["string", "null"]
        customLocationName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        city:
          type: string
        capacity:
          type: integer
        lengthFt:
          type: ["string", "null"]
        yearBuilt:
          type: ["integer", "null"]
        manufacturer:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
        status:
          type: string
        isFeatured:
          type: boolean
        featuredSortOrder:
          type: integer
        featuredFrom:
          type: ["string", "null"]
        featuredUntil:
          type: ["string", "null"]
        avgRating:
          type: string
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - hostId
        - title
        - location
        - city
        - capacity
        - status
        - isFeatured
        - featuredSortOrder
        - avgRating
        - reviewCount
        - createdAt
        - updatedAt
    YachtDetail:
      type: object
      properties:
        yacht:
          $ref: "#/components/schemas/Yacht"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/YachtPhoto"
        pricing:
          type: array
          items:
            $ref: "#/components/schemas/YachtPricing"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        host:
          $ref: "#/components/schemas/HostPublic"
      required:
        - yacht
        - photos
        - pricing
        - reviews
    YachtPhoto:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - url
        - isPrimary
        - sortOrder
    YachtPricing:
      type: object
      properties:
        templateId:
          type: string
        templateName:
          type: string
        durationHours:
          type: integer
        priceEgp:
          type: string
      required:
        - templateId
        - templateName
        - durationHours
        - priceEgp
    YachtListResponse:
      type: object
      properties:
        yachts:
          type: array
          items:
            $ref: "#/components/schemas/Yacht"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - yachts
        - total
        - page
        - limit
...
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: string
        yachtId:
          type: string
        templateId:
          type: string
        date:
          type: string
        startTime:
          type: string
        isAvailable:
          type: boolean
        priceOverrideEgp:
          type: ["string", "null"]
        effectivePriceEgp:
          type: ["string", "null"]
        displayStatus:
          type: string
          enum: [available, blocked, held, booked, past]
        editable:
          type: boolean
        bookingId:
          type: ["string", "null"]
      required:
        - id
        - yachtId
        - templateId
        - date
        - startTime
        - isAvailable
        - effectivePriceEgp
        - displayStatus
        - editable
    AvailabilityListResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlot"
      required:
        - slots
...
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        reviewerId:
          type: string
        revieweeId:
          type: string
        yachtId:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - reviewerId
        - revieweeId
        - rating
        - type
        - status
        - createdAt
...
    HostPublic:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        memberSince:
          type: string
      required:
        - id
        - memberSince
...